Move isatty and get_blocking to mock termios lib.
